fix: Renamed transaction header keys
