+++ client/src/main/java/com/tigrisdata/db/client/StandardTigrisAsyncClient.java
@@ -18,10 +18,10 @@ import com.google.common.annotations.VisibleForTesting;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.tigrisdata.db.api.v1.grpc.Api;
 import com.tigrisdata.db.api.v1.grpc.TigrisGrpc;
-import static com.tigrisdata.db.client.Messages.CREATE_DB_FAILED;
-import static com.tigrisdata.db.client.Messages.DROP_DB_FAILED;
-import static com.tigrisdata.db.client.Messages.LIST_DBS_FAILED;
-import static com.tigrisdata.db.client.Messages.SERVER_METADATA_FAILED;
+import static com.tigrisdata.db.client.Constants.CREATE_DB_FAILED;
+import static com.tigrisdata.db.client.Constants.DROP_DB_FAILED;
+import static com.tigrisdata.db.client.Constants.LIST_DBS_FAILED;
+import static com.tigrisdata.db.client.Constants.SERVER_METADATA_FAILED;
 import static com.tigrisdata.db.client.TypeConverter.toCreateDatabaseRequest;
 import static com.tigrisdata.db.client.TypeConverter.toDropDatabaseRequest;
 import static com.tigrisdata.db.client.TypeConverter.toListDatabasesRequest;
