+++ client/src/main/java/com/tigrisdata/db/client/StandardTigrisAsyncCollection.java
@@ -18,13 +18,13 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.tigrisdata.db.api.v1.grpc.Api;
 import com.tigrisdata.db.api.v1.grpc.TigrisGrpc;
-import static com.tigrisdata.db.client.Messages.DELETE_FAILED;
-import static com.tigrisdata.db.client.Messages.DESCRIBE_COLLECTION_FAILED;
-import static com.tigrisdata.db.client.Messages.INSERT_FAILED;
-import static com.tigrisdata.db.client.Messages.INSERT_OR_REPLACE_FAILED;
-import static com.tigrisdata.db.client.Messages.JSON_SER_DE_ERROR;
-import static com.tigrisdata.db.client.Messages.READ_FAILED;
-import static com.tigrisdata.db.client.Messages.UPDATE_FAILED;
+import static com.tigrisdata.db.client.Constants.DELETE_FAILED;
+import static com.tigrisdata.db.client.Constants.DESCRIBE_COLLECTION_FAILED;
+import static com.tigrisdata.db.client.Constants.INSERT_FAILED;
+import static com.tigrisdata.db.client.Constants.INSERT_OR_REPLACE_FAILED;
+import static com.tigrisdata.db.client.Constants.JSON_SER_DE_ERROR;
+import static com.tigrisdata.db.client.Constants.READ_FAILED;
+import static com.tigrisdata.db.client.Constants.UPDATE_FAILED;
 import static com.tigrisdata.db.client.TypeConverter.extractTigrisError;
 import static com.tigrisdata.db.client.TypeConverter.makeTransactionAware;
 import static com.tigrisdata.db.client.TypeConverter.readOneDefaultReadRequestOptions;
