+++ client/src/main/java/com/tigrisdata/db/client/StandardTigrisAsyncDatabase.java
@@ -17,12 +17,12 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.tigrisdata.db.api.v1.grpc.Api;
 import com.tigrisdata.db.api.v1.grpc.TigrisGrpc;
-import static com.tigrisdata.db.client.Messages.BEGIN_TRANSACTION_FAILED;
-import static com.tigrisdata.db.client.Messages.DESCRIBE_DB_FAILED;
-import static com.tigrisdata.db.client.Messages.DROP_COLLECTION_FAILED;
-import static com.tigrisdata.db.client.Messages.LIST_COLLECTION_FAILED;
-import static com.tigrisdata.db.client.Messages.STREAM_CONVERT_FAILED;
-import static com.tigrisdata.db.client.Messages.STREAM_FAILED;
+import static com.tigrisdata.db.client.Constants.BEGIN_TRANSACTION_FAILED;
+import static com.tigrisdata.db.client.Constants.DESCRIBE_DB_FAILED;
+import static com.tigrisdata.db.client.Constants.DROP_COLLECTION_FAILED;
+import static com.tigrisdata.db.client.Constants.LIST_COLLECTION_FAILED;
+import static com.tigrisdata.db.client.Constants.STREAM_CONVERT_FAILED;
+import static com.tigrisdata.db.client.Constants.STREAM_FAILED;
 import static com.tigrisdata.db.client.TypeConverter.toBeginTransactionRequest;
 import static com.tigrisdata.db.client.TypeConverter.toDatabaseDescription;
 import static com.tigrisdata.db.client.TypeConverter.toDropCollectionRequest;
