+++ client/src/main/java/com/tigrisdata/db/client/StandardTigrisDatabase.java
@@ -16,13 +16,13 @@ package com.tigrisdata.db.client;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.tigrisdata.db.api.v1.grpc.Api;
 import com.tigrisdata.db.api.v1.grpc.TigrisGrpc;
-import static com.tigrisdata.db.client.Messages.BEGIN_TRANSACTION_FAILED;
-import static com.tigrisdata.db.client.Messages.CREATE_OR_UPDATE_COLLECTION_FAILED;
-import static com.tigrisdata.db.client.Messages.DESCRIBE_DB_FAILED;
-import static com.tigrisdata.db.client.Messages.DROP_COLLECTION_FAILED;
-import static com.tigrisdata.db.client.Messages.LIST_COLLECTION_FAILED;
-import static com.tigrisdata.db.client.Messages.STREAM_FAILED;
-import static com.tigrisdata.db.client.Messages.TRANSACTION_FAILED;
+import static com.tigrisdata.db.client.Constants.BEGIN_TRANSACTION_FAILED;
+import static com.tigrisdata.db.client.Constants.CREATE_OR_UPDATE_COLLECTION_FAILED;
+import static com.tigrisdata.db.client.Constants.DESCRIBE_DB_FAILED;
+import static com.tigrisdata.db.client.Constants.DROP_COLLECTION_FAILED;
+import static com.tigrisdata.db.client.Constants.LIST_COLLECTION_FAILED;
+import static com.tigrisdata.db.client.Constants.STREAM_FAILED;
+import static com.tigrisdata.db.client.Constants.TRANSACTION_FAILED;
 import com.tigrisdata.db.client.error.TigrisException;
 import com.tigrisdata.db.type.TigrisCollectionType;
 import com.tigrisdata.tools.schema.core.ModelToJsonSchema;
